requirement.txt
streamlit
pandas
mysql-connector-python
datetime
random
csv